// src/config/interfaces.ts
export interface Pph21StateInterface {
    personal: {
        type: string;
        name: string;
        label: string;
        value: string;
        classNames: {
            inputWrapper: string;
        };
        color?: string;
        itemClassname?: string;
        items?: any[];
    }[];
    earnings: {
        type: string;
        name: string;
        label: string;
        value: string;
        classNames: {
            inputWrapper: string;
        };
        color?: string;
        itemClassname?: string;
        items?: any[];
    }[];
    deductions: {
        type: string;
        name: string;
        label: string;
        value: string;
        classNames: {
            inputWrapper: string;
        };
        color?: string;
        itemClassname?: string;
        items?: any[];
    }[];
    result: {
        type: string;
        name: string;
        label: string;
        value: string;
        classNames: {
            inputWrapper: string;
        };
        color?: string;
        itemClassname?: string;
        items?: any[];
    }[];
}

export interface Pph21TerStateInterface {
    personal: {
        type: string;
        name: string;
        label: string;
        value: string;
        classNames: {
            inputWrapper: string;
        };
        color?: string;
        itemClassname?: string;
        items?: any[];
    }[];
    earnings: {
        type: string;
        name: string;
        label: string;
        value: string;
        classNames: {
            inputWrapper: string;
        };
        color?: string;
        itemClassname?: string;
        items?: any[];
    }[];
    result: {
        type: string;
        name: string;
        label: string;
        value: string;
        classNames: {
            inputWrapper: string;
        };
        color?: string;
        itemClassname?: string;
        items?: any[];
    }[];
}

export interface PpnStateInterface {
    settings: {
        type: string;
        name: string;
        label: string;
        value: string;
        classNames: {
            inputWrapper: string;
        };
        color?: string;
        itemClassname?: string;
        items?: any[];
    }[];
    components: {
        type: string;
        name: string;
        label: string;
        value: string;
        classNames: {
            inputWrapper: string;
        };
        color?: string;
        itemClassname?: string;
        items?: any[];
    }[];
    result: {
        type: string;
        name: string;
        label: string;
        value: string;
        classNames: {
            inputWrapper: string;
        };
        color?: string;
        itemClassname?: string;
        items?: any[];
    }[];
}